requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==22.0.0
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.12.0
whitenoise==6.6.0
